project_management/requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
python-dotenv==1.0.0
Werkzeug==2.3.6
WTForms==3.0.1
email-validator==2.1.1
pymysql==1.1.0
Flask-Mail==0.9.1
Flask-Babel==3.1.0
Pillow==10.0.0
pytz==2023.3
markdown==3.4.4
gunicorn==21.2.0
